timeSlot fixed

+++ app/src/main/java/com/polito/timebanking/models/User.kt
@@ -4,10 +4,10 @@ import androidx.room.Entity
 import androidx.room.Index
 import androidx.room.PrimaryKey
 
-
 @Entity(indices = [Index(value = ["email"], unique = true)])
 data class User(
-    @PrimaryKey(autoGenerate = true) val id: Long,
+    @PrimaryKey(autoGenerate = true)
+    val id: Int,
     var fullName: String,
     var nickname: String,
     var email: String,
