+++ app/src/main/java/com/polito/timebanking/models/UserSkillDao.kt
@@ -13,7 +13,7 @@ interface UserSkillDao {
                 "INNER JOIN UserSkill ON UserSkill.skillId = Skill.id " +
                 "WHERE UserSkill.userId = :userId"
     )
-    fun getUserSkills(userId: Long): LiveData<List<Skill>>
+    fun getUserSkills(userId: Int): LiveData<List<Skill>>
 
     @Insert
     fun insertUserSkill(userSkill: UserSkill)
