+++ app/src/main/java/com/polito/timebanking/models/UserSkillRepository.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.LiveData
 class UserSkillRepository(application: Application) {
     private val userSkillDao = TimeBankingDatabase.getDatabase(application).userSkillDao()
 
-    fun getUserSkills(userId: Long): LiveData<List<Skill>> = userSkillDao.getUserSkills(userId)
+    fun getUserSkills(userId: Int): LiveData<List<Skill>> = userSkillDao.getUserSkills(userId)
 
     fun insertUserSkill(userSkill: UserSkill) {
         userSkillDao.insertUserSkill(userSkill)
